refactor: review types 파일 위치변경
