+++ src/reviews/review.controller.ts
@@ -3,9 +3,15 @@ import { IReviewController } from '#reviews/interfaces/review.controller.interfa
 import { Body, Controller, Delete, Get, HttpStatus, Param, Patch, Post, Query, UseGuards } from '@nestjs/common';
 import { ApiBearerAuth, ApiBody, ApiOperation, ApiParam, ApiResponse } from '@nestjs/swagger';
 import { ReviewService } from './review.service.js';
-import { DriverReviewResponseDTO, MyReviewResponseDTO, ReviewBodyDTO, ReviewOutputDTO, PatchReviewDTO } from './review.types.js';
 import { GetQueries } from '#types/queries.type.js';
 import { SortOrder } from '#types/options.type.js';
+import {
+  DriverReviewResponseDTO,
+  MyReviewResponseDTO,
+  PatchReviewDTO,
+  ReviewBodyDTO,
+  ReviewOutputDTO,
+} from './types/review.dto.js';
 
 @Controller('reviews')
 export class ReviewController implements IReviewController {
